sep: server and location_block.cpp
