+++ Makefile
@@ -19,6 +19,8 @@ CONFIG_DIR	=	Config
 SRCS		+=	$(CONFIG_DIR)/FileHandler/FileHandler.cpp \
 				$(CONFIG_DIR)/ConfigParser/ConfigParser.cpp \
 				$(CONFIG_DIR)/ConfigParser/http_block.cpp \
+				$(CONFIG_DIR)/ConfigParser/server_block.cpp \
+				$(CONFIG_DIR)/ConfigParser/location_block.cpp \
 				$(CONFIG_DIR)/ConfigParser/parse_timeout.cpp \
 				$(CONFIG_DIR)/Token/Token.cpp \
 				$(CONFIG_DIR)/Tokenizer/Tokenizer.cpp \
