+++ srcs/Config/ConfigParser/error_msg.cpp
@@ -1,0 +1,10 @@
+#include <deque>
+#include <limits>
+#include <set>
+#include <utility>
+#include "Constant.hpp"
+#include "FileHandler.hpp"
+#include "ConfigParser.hpp"
+#include "Tokenizer.hpp"
+#include "StringHandler.hpp"
+#include "HttpMessageParser.hpp"
